Migrate pending approval into a component
