+++ flexget/tests/test_pending_approval.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals, division, absolute_import
 from builtins import *  # noqa pylint: disable=unused-import, redefined-builtin
 
 from flexget.manager import Session
-from flexget.plugins.filter.pending_approval import PendingEntry
+from flexget.components.pending_approval.db import PendingEntry
 
 
 class TestPendingApproval(object):
